Compiles and let DMET iterations continue again.

+++ FCI.cpp
@@ -304,7 +304,7 @@ short int CountSameImpurity(std::vector<bool> BraString, std::vector<bool> KetSt
 }
 
 /* This calculates the elements < a_i,alpha^\dagger a_j,alpha + a_i,beta^\dagger a_j,beta > */
-Eigen::MatrixXd Form1RDM(InputObj &Input, int FragmentIndex, Eigen::VectorXd Eigenvector, std::vector< std::vector< bool > > aStrings, std::vector< std::vector< bool > > bStrings)
+Eigen::MatrixXd Form1RDM(InputObj &Input, int FragmentIndex, Eigen::VectorXf Eigenvector, std::vector< std::vector< bool > > aStrings, std::vector< std::vector< bool > > bStrings)
 {
     std::vector<int> FragPos;
     std::vector<int> BathPos;
